Refactor in-memory storage id generator
